Divide components
