+++ src/components/About.tsx
@@ -1,0 +1,15 @@
+import * as React from 'react';
+
+export class About extends React.Component<{}, undefined> {
+    constructor(props: {}) {
+        super(props);
+    }
+
+    public render() {
+        return (
+            <div>
+                <h2>About</h2>
+            </div>
+        );
+    }
+}
