+++ src/components/Home.tsx
@@ -1,0 +1,15 @@
+import * as React from 'react';
+
+export class Home extends React.Component<{}, undefined> {
+    constructor(props: {}) {
+        super(props);
+    }
+
+    public render() {
+        return (
+            <div>
+                <h2>Home</h2>
+            </div>
+        );
+    }
+}
